# [Visual] VisualModelImpl: Clean and optimize slightly ComputeNormals()