- clean (const, foreach loop)
- optimize a little bit when initializing the normals vector
   - if there is not index, use memset to clear all vectors at once
   - if there are indices, just dont clear the new vectors as there are supposed to be 0 already.

On a big mesh (315k vertices and 620k triangles) and mapped (to force the call to computeNormals)
 - before  `1000 iterations done in 14.95 s ( 66.8896 FPS).`
 - after `1000 iterations done in 14.075 s ( 71.048 FPS)`

The case is not really usual as it needs huge meshes to see the difference ; and mappings will certainly takes much more time


______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).